Remove old editors for kie-deployment-descriptor.xml
