--- jbpm-wb-integration/jbpm-wb-integration-api/src/main/java/org/jbpm/workbench/wi/dd/type/DDResourceTypeDefinition.java
@@ -1,61 +1,0 @@
-/*
- * Copyright 2017 Red Hat, Inc. and/or its affiliates.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.jbpm.workbench.wi.dd.type;
-
-import javax.enterprise.context.ApplicationScoped;
-
-import org.uberfire.backend.vfs.Path;
-import org.uberfire.workbench.type.ResourceTypeDefinition;
-
-@ApplicationScoped
-public class DDResourceTypeDefinition implements ResourceTypeDefinition {
-
-    @Override
-    public String getShortName() {
-        return "kie deployment descriptor";
-    }
-
-    @Override
-    public String getDescription() {
-        return null;
-    }
-
-    @Override
-    public String getPrefix() {
-        return "kie-deployment-descriptor";
-    }
-
-    @Override
-    public String getSuffix() {
-        return "xml";
-    }
-
-    @Override
-    public int getPriority() {
-        return 0;
-    }
-
-    @Override
-    public String getSimpleWildcardPattern() {
-        return getPrefix() + "." + getSuffix();
-    }
-
-    @Override
-    public boolean accept(final Path path) {
-        return path.getFileName().equals(getPrefix() + "." + getSuffix());
-    }
-}
