--- jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/DDParametersPopup.ui.xml
@@ -1,55 +1,0 @@
-<!--
-  ~ Copyright 2017 Red Hat, Inc. and/or its affiliates.
-  ~
-  ~ Licensed under the Apache License, Version 2.0 (the "License");
-  ~ you may not use this file except in compliance with the License.
-  ~ You may obtain a copy of the License at
-  ~
-  ~       http://www.apache.org/licenses/LICENSE-2.0
-  ~
-  ~ Unless required by applicable law or agreed to in writing, software
-  ~ distributed under the License is distributed on an "AS IS" BASIS,
-  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-  ~ See the License for the specific language governing permissions and
-  ~ limitations under the License.
-  -->
-
-<!DOCTYPE ui:UiBinder SYSTEM "http://dl.google.com/gwt/DTD/xhtml.ent">
-<ui:UiBinder xmlns:ui="urn:ui:com.google.gwt.uibinder"
-             xmlns:g="urn:import:com.google.gwt.user.client.ui"
-             xmlns:b="urn:import:org.gwtbootstrap3.client.ui"
-             xmlns:bhtml="urn:import:org.gwtbootstrap3.client.ui.html"
-             xmlns:bgwt="urn:import:org.gwtbootstrap3.client.ui.gwt">
-
-  <ui:with field="i18n" type="org.jbpm.workbench.wi.client.i18n.Constants"/>
-  <ui:style>
-    .edit-cell input {
-      width: 95%;
-      text-align: left;
-    }
-  </ui:style>
-  <g:HTMLPanel>
-      <b:FieldSet>
-        <b:Legend>
-          <ui:text from="{i18n.DeploymentDescriptorParameters}"/>
-        </b:Legend>
-        <b:Alert type="INFO" dismissable="false">
-          <b:Description>
-            <b:DescriptionTitle text="{i18n.NamedParams}"/>
-            <b:DescriptionData text="ksession"/>
-            <b:DescriptionData text="taskService"/>
-            <b:DescriptionData text="runtimeManager"/>
-            <b:DescriptionData text="classLoader"/>
-            <b:DescriptionData text="entityManagerFactory"/>
-          </b:Description>
-        </b:Alert>
-        <b:FormGroup ui:field="parametersGroup">
-          <b:FormLabel for="ou" ui:field="parametersLabel"/>
-          <bgwt:CellTable ui:field="parametersTable" addStyleNames='{style.edit-cell}'/>
-          <b:HelpBlock ui:field="parametersHelpInline"/>
-          <b:Button ui:field="addParameterButton" icon="PLUS" type="PRIMARY" text="{i18n.Add}"/>
-        </b:FormGroup>
-      </b:FieldSet>
-  </g:HTMLPanel>
-
-</ui:UiBinder>
