+++ jbpm-wb-integration/jbpm-wb-integration-client/src/main/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/items/ParametersModalView.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.items;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptor.items;
 
 import javax.inject.Inject;
 import javax.inject.Named;
