+++ jbpm-wb-integration/jbpm-wb-integration-client/src/test/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/items/NamedObjectItemPresenterTest.java
@@ -14,12 +14,11 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.items;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptor.items;
 
 import com.google.common.collect.ImmutableList;
-import org.jbpm.workbench.wi.client.editors.deployment.descriptornew.DeploymentsSectionPresenter;
-import org.jbpm.workbench.wi.client.editors.deployment.descriptornew.DeploymentsSectionPresenter.MarshallingStrategiesListPresenter;
-import org.jbpm.workbench.wi.client.editors.deployment.descriptornew.model.Resolver;
+import org.jbpm.workbench.wi.client.editors.deployment.descriptor.DeploymentsSectionPresenter;
+import org.jbpm.workbench.wi.client.editors.deployment.descriptor.model.Resolver;
 import org.jbpm.workbench.wi.dd.model.ItemObjectModel;
 import org.jbpm.workbench.wi.dd.model.Parameter;
 import org.junit.Before;
@@ -37,10 +36,10 @@ import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 
 @RunWith(MockitoJUnitRunner.class)
-public class ObjectItemPresenterTest {
+public class NamedObjectItemPresenterTest {
 
     @Mock
-    private ObjectItemPresenter.View view;
+    private NamedObjectItemPresenter.View view;
 
     @Mock
     private ParametersModal parametersModal;
@@ -48,50 +47,51 @@ public class ObjectItemPresenterTest {
     @Mock
     private KieEnumSelectElement<Resolver> resolversSelect;
 
-    private ObjectItemPresenter objectItemPresenter;
+    private NamedObjectItemPresenter namedObjectItemPresenter;
 
     @Before
     public void before() {
-        objectItemPresenter = spy(new ObjectItemPresenter(view,
-                                                          parametersModal,
-                                                          resolversSelect));
+        namedObjectItemPresenter = spy(new NamedObjectItemPresenter(view,
+                                                                    parametersModal,
+                                                                    resolversSelect));
     }
 
     @Test
     public void testSetup() {
-        final ItemObjectModel model = spy(new ItemObjectModel(null, "Value", "reflection", ImmutableList.of(new Parameter("Foo", "Bar"))));
+        final ItemObjectModel model = spy(new ItemObjectModel("Name", "Value", "mvel", ImmutableList.of(new Parameter("Foo", "Bar"))));
 
-        objectItemPresenter.setup(model, mock(DeploymentsSectionPresenter.class));
+        namedObjectItemPresenter.setup(model, mock(DeploymentsSectionPresenter.class));
 
         verify(model, never()).setParameters(any());
-        verify(view).init(eq(objectItemPresenter));
+        verify(view).init(eq(namedObjectItemPresenter));
         verify(view).setValue(eq("Value"));
+        verify(view).setName(eq("Name"));
         verify(view).setParametersCount(eq(1));
         verify(parametersModal).setup(any(), any());
-        verify(resolversSelect).setup(any(), any(), eq(Resolver.REFLECTION), any());
+        verify(resolversSelect).setup(any(), any(), eq(Resolver.MVEL), any());
     }
 
     @Test
     public void testRemove() {
         final DeploymentsSectionPresenter parentPresenter = mock(DeploymentsSectionPresenter.class);
-        final MarshallingStrategiesListPresenter listPresenter = mock(MarshallingStrategiesListPresenter.class);
+        final DeploymentsSectionPresenter.MarshallingStrategiesListPresenter listPresenter = mock(DeploymentsSectionPresenter.MarshallingStrategiesListPresenter.class);
 
-        objectItemPresenter.parentPresenter = parentPresenter;
-        objectItemPresenter.setListPresenter(listPresenter);
+        namedObjectItemPresenter.parentPresenter = parentPresenter;
+        namedObjectItemPresenter.setListPresenter(listPresenter);
 
-        objectItemPresenter.remove();
+        namedObjectItemPresenter.remove();
 
-        verify(listPresenter).remove(eq(objectItemPresenter));
+        verify(listPresenter).remove(eq(namedObjectItemPresenter));
         verify(parentPresenter).fireChangeEvent();
     }
 
     @Test
     public void testSignalParameterAddedOrRemoved() {
         final DeploymentsSectionPresenter parentPresenter = mock(DeploymentsSectionPresenter.class);
-        objectItemPresenter.parentPresenter = parentPresenter;
-        objectItemPresenter.model = new ItemObjectModel(null, "Value", "reflection", ImmutableList.of(new Parameter("Foo", "Bar")));
+        namedObjectItemPresenter.parentPresenter = parentPresenter;
+        namedObjectItemPresenter.model = new ItemObjectModel("Name", "Value", "mvel", ImmutableList.of(new Parameter("Foo", "Bar")));
 
-        objectItemPresenter.signalParameterAddedOrRemoved();
+        namedObjectItemPresenter.signalParameterAddedOrRemoved();
 
         verify(view).setParametersCount(eq(1));
         verify(parentPresenter).fireChangeEvent();
