+++ jbpm-wb-integration/jbpm-wb-integration-client/src/test/java/org/jbpm/workbench/wi/client/editors/deployment/descriptor/items/RequiredRolesListItemPresenterTest.java
@@ -14,10 +14,10 @@
  * limitations under the License.
  */
 
-package org.jbpm.workbench.wi.client.editors.deployment.descriptornew.items;
+package org.jbpm.workbench.wi.client.editors.deployment.descriptor.items;
 
-import org.jbpm.workbench.wi.client.editors.deployment.descriptornew.DeploymentsSectionPresenter;
-import org.jbpm.workbench.wi.client.editors.deployment.descriptornew.DeploymentsSectionPresenter.RequiredRolesListPresenter;
+import org.jbpm.workbench.wi.client.editors.deployment.descriptor.DeploymentsSectionPresenter;
+import org.jbpm.workbench.wi.client.editors.deployment.descriptor.DeploymentsSectionPresenter.RequiredRolesListPresenter;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
